pyproject.toml
[tool.poetry]
name = "lanyard-langserve"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
uvicorn = "^0.23.2"
torch="2.2.1"
transformers="4.44.2"
scikit-learn="1.5.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = ">=2.7.4,<3.0.0"
python-dotenv = "^1.0.1"
langchain-google-genai = "^1.0.8"
langchain-pinecone = "^0.1.3"
langchain = "^0.2.12"
langchain-community = "^0.2.11"
psycopg2-binary = "^2.9.9"
beautifulsoup4 = "^4.12.3"
jq = "^1.7.0"
langchain-groq = "^0.1.9"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
